# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Golf caddie backend — players, rounds, shots and course data.",
        "title": "Agentic Caddie API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/tees": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tees"
                ],
                "summary": "List all tees",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.teeResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tees"
                ],
                "summary": "Create a tee",
                "parameters": [
                    {
                        "description": "Tee to create",
                        "name": "tee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createTeeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.teeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tees/course/{courseId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tees"
                ],
                "summary": "List tees for a course",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.teeResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid course ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tees/course/{courseId}/{name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tees"
                ],
                "summary": "Get a tee by course and name",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tee name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.teeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid course ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Tee not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tees/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tees"
                ],
                "summary": "Get a tee by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.teeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid tee ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Tee not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "tees"
                ],
                "summary": "Update a tee's ratings",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated ratings",
                        "name": "tee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateTeeRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "tees"
                ],
                "summary": "Delete a tee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid tee ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.createTeeRequest": {
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "integer",
                    "example": 1
                },
                "course_rating": {
                    "type": "number",
                    "example": 69.5
                },
                "name": {
                    "type": "string",
                    "example": "white"
                },
                "slope_rating": {
                    "type": "integer",
                    "example": 113
                }
            }
        },
...
        "handlers.teeResponse": {
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "integer"
                },
                "course_rating": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slope_rating": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.updateTeeRequest": {
            "type": "object",
            "properties": {
                "course_rating": {
                    "type": "number",
                    "example": 69.5
                },
                "slope_rating": {
                    "type": "integer",
                    "example": 113
                }
            }
        }
    }
}